Revert "Embedded competing"
