+++ src/EventStore.ClientAPI.Embedded/EmbeddedResponders.cs
@@ -9,8 +9,7 @@ namespace EventStore.ClientAPI.Embedded
 {
     internal static class EmbeddedResponders
     {
-        internal class AppendToStream :
-            EmbeddedResponderBase<WriteResult, ClientMessage.WriteEventsCompleted>
+        internal class AppendToStream : EmbeddedResponderBase<WriteResult, ClientMessage.WriteEventsCompleted>
         {
             private readonly int _expectedVersion;
             private readonly string _stream;
@@ -110,8 +109,7 @@ namespace EventStore.ClientAPI.Embedded
             }
         }
 
-        internal class ReadAllEventsBackward :
-            EmbeddedResponderBase<AllEventsSlice, ClientMessage.ReadAllEventsBackwardCompleted>
+        internal class ReadAllEventsBackward : EmbeddedResponderBase<AllEventsSlice, ClientMessage.ReadAllEventsBackwardCompleted>
         {
             public ReadAllEventsBackward(TaskCompletionSource<AllEventsSlice> source)
                 : base(source)
@@ -185,8 +183,7 @@ namespace EventStore.ClientAPI.Embedded
             }
         }
 
-        internal class ReadEvent : 
-            EmbeddedResponderBase<EventReadResult, ClientMessage.ReadEventCompleted>
+        internal class ReadEvent : EmbeddedResponderBase<EventReadResult, ClientMessage.ReadEventCompleted>
         {
             private readonly int _eventNumber;
             private readonly string _stream;
@@ -243,8 +240,7 @@ namespace EventStore.ClientAPI.Embedded
             }
         }
 
-        internal class ReadStreamEventsBackward : 
-            EmbeddedResponderBase<StreamEventsSlice, ClientMessage.ReadStreamEventsBackwardCompleted>
+        internal class ReadStreamEventsBackward : EmbeddedResponderBase<StreamEventsSlice, ClientMessage.ReadStreamEventsBackwardCompleted>
         {
             private readonly int _fromEventNumber;
             private readonly string _stream;
@@ -305,8 +301,7 @@ namespace EventStore.ClientAPI.Embedded
             }
         }
 
-        internal class ReadStreamForwardEvents : 
-            EmbeddedResponderBase<StreamEventsSlice, ClientMessage.ReadStreamEventsForwardCompleted>
+        internal class ReadStreamForwardEvents : EmbeddedResponderBase<StreamEventsSlice, ClientMessage.ReadStreamEventsForwardCompleted>
         {
             private readonly int _fromEventNumber;
             private readonly string _stream;
@@ -366,7 +361,7 @@ namespace EventStore.ClientAPI.Embedded
             }
         }
 
-        internal class TransactionCommit :
+        public class TransactionCommit :
             EmbeddedResponderBase<WriteResult, ClientMessage.TransactionCommitCompleted>
         {
             public TransactionCommit(TaskCompletionSource<WriteResult> source)
@@ -410,7 +405,7 @@ namespace EventStore.ClientAPI.Embedded
             }
         }
 
-        internal class TransactionStart :
+        public class TransactionStart :
             EmbeddedResponderBase<EventStoreTransaction, ClientMessage.TransactionStartCompleted>
         {
             private readonly int _expectedVersion;
@@ -459,8 +454,8 @@ namespace EventStore.ClientAPI.Embedded
             }
         }
 
-        internal class TransactionWrite :
-            EmbeddedResponderBase<EventStoreTransaction, ClientMessage.TransactionWriteCompleted>
+        public class TransactionWrite :
+        EmbeddedResponderBase<EventStoreTransaction, ClientMessage.TransactionWriteCompleted>
         {
             private readonly IEventStoreTransactionConnection _parentConnection;
 
@@ -492,126 +487,6 @@ namespace EventStore.ClientAPI.Embedded
             protected override EventStoreTransaction TransformResponse(ClientMessage.TransactionWriteCompleted response)
             {
                 return new EventStoreTransaction(response.TransactionId, null, _parentConnection);
-            }
-        }
-
-        internal class CreatePersistentSubscription : 
-            EmbeddedResponderBase<PersistentSubscriptionCreateResult, ClientMessage.CreatePersistentSubscriptionCompleted> 
-        {
-            private readonly string _stream;
-            private readonly string _groupName;
-
-            public CreatePersistentSubscription(TaskCompletionSource<PersistentSubscriptionCreateResult> source, string stream, string groupName) 
-                : base(source)
-            {
-                _groupName = groupName;
-                _stream = stream;
-            }
-
-            protected override void InspectResponse(ClientMessage.CreatePersistentSubscriptionCompleted response)
-            {
-                switch (response.Result)
-                {
-                    case ClientMessage.CreatePersistentSubscriptionCompleted.CreatePersistentSubscriptionResult.Success:
-                        Succeed(response);
-                        break;
-                    case ClientMessage.CreatePersistentSubscriptionCompleted.CreatePersistentSubscriptionResult.Fail:
-                        Fail(new InvalidOperationException(String.Format("Subscription group {0} on stream {1} failed '{2}'", _groupName, _stream, response.Reason)));
-                        break;
-                    case ClientMessage.CreatePersistentSubscriptionCompleted.CreatePersistentSubscriptionResult.AccessDenied:
-                        Fail(new AccessDeniedException(string.Format("Write access denied for stream '{0}'.", _stream)));
-                        break;
-                    case ClientMessage.CreatePersistentSubscriptionCompleted.CreatePersistentSubscriptionResult.AlreadyExists:
-                        Fail(new InvalidOperationException(String.Format("Subscription group {0} on stream {1} alreay exists", _groupName, _stream)));
-                        break;
-                    default:
-                        throw new Exception(string.Format("Unexpected OperationResult: {0}.", response.Result));
-                }
-            }
-
-            protected override PersistentSubscriptionCreateResult TransformResponse(ClientMessage.CreatePersistentSubscriptionCompleted response)
-            {
-                return new PersistentSubscriptionCreateResult((PersistentSubscriptionCreateStatus)response.Result);
-            }
-        }
-
-        internal class UpdatePersistentSubscription :
-            EmbeddedResponderBase<PersistentSubscriptionUpdateResult, ClientMessage.UpdatePersistentSubscriptionCompleted> 
-        {
-            private readonly string _stream;
-            private readonly string _groupName;
-
-            public UpdatePersistentSubscription(TaskCompletionSource<PersistentSubscriptionUpdateResult> source, string stream, string groupName) 
-                : base(source)
-            {
-                _groupName = groupName;
-                _stream = stream;
-            }
-
-            protected override void InspectResponse(ClientMessage.UpdatePersistentSubscriptionCompleted response)
-            {
-                switch (response.Result)
-                {
-                    case ClientMessage.UpdatePersistentSubscriptionCompleted.UpdatePersistentSubscriptionResult.Success:
-                        Succeed(response);
-                        break;
-                    case ClientMessage.UpdatePersistentSubscriptionCompleted.UpdatePersistentSubscriptionResult.Fail:
-                        Fail(new InvalidOperationException(String.Format("Subscription group {0} on stream {1} failed '{2}'", _groupName, _stream, response.Reason)));
-                        break;
-                    case ClientMessage.UpdatePersistentSubscriptionCompleted.UpdatePersistentSubscriptionResult.AccessDenied:
-                        Fail(new AccessDeniedException(string.Format("Write access denied for stream '{0}'.", _stream)));
-                        break;
-                    case ClientMessage.UpdatePersistentSubscriptionCompleted.UpdatePersistentSubscriptionResult.DoesNotExist:
-                        Fail(new InvalidOperationException(String.Format("Subscription group {0} on stream {1} does not exist", _groupName, _stream)));
-                        break;
-                    default:
-                        throw new Exception(string.Format("Unexpected OperationResult: {0}.", response.Result));
-                }
-            }
-
-            protected override PersistentSubscriptionUpdateResult TransformResponse(ClientMessage.UpdatePersistentSubscriptionCompleted response)
-            {
-                return new PersistentSubscriptionUpdateResult((PersistentSubscriptionUpdateStatus)response.Result);
-            }
-        }
-
-        internal class DeletePersistentSubscription :
-            EmbeddedResponderBase<PersistentSubscriptionDeleteResult, ClientMessage.DeletePersistentSubscriptionCompleted> 
-        {
-            private readonly string _stream;
-            private readonly string _groupName;
-
-            public DeletePersistentSubscription(TaskCompletionSource<PersistentSubscriptionDeleteResult> source, string stream, string groupName) 
-                : base(source)
-            {
-                _groupName = groupName;
-                _stream = stream;
-            }
-
-            protected override void InspectResponse(ClientMessage.DeletePersistentSubscriptionCompleted response)
-            {
-                switch (response.Result)
-                {
-                    case ClientMessage.DeletePersistentSubscriptionCompleted.DeletePersistentSubscriptionResult.Success:
-                        Succeed(response);
-                        break;
-                    case ClientMessage.DeletePersistentSubscriptionCompleted.DeletePersistentSubscriptionResult.Fail:
-                        Fail(new InvalidOperationException(String.Format("Subscription group {0} on stream {1} failed '{2}'", _groupName, _stream, response.Reason)));
-                        break;
-                    case ClientMessage.DeletePersistentSubscriptionCompleted.DeletePersistentSubscriptionResult.AccessDenied:
-                        Fail(new AccessDeniedException(string.Format("Write access denied for stream '{0}'.", _stream)));
-                        break;
-                    case ClientMessage.DeletePersistentSubscriptionCompleted.DeletePersistentSubscriptionResult.DoesNotExist:
-                        Fail(new InvalidOperationException(String.Format("Subscription group {0} on stream {1} does not exist", _groupName, _stream)));
-                        break;
-                    default:
-                        throw new Exception(string.Format("Unexpected OperationResult: {0}.", response.Result));
-                }
-            }
-
-            protected override PersistentSubscriptionDeleteResult TransformResponse(ClientMessage.DeletePersistentSubscriptionCompleted response)
-            {
-                return new PersistentSubscriptionDeleteResult((PersistentSubscriptionDeleteStatus)response.Result);
             }
         }
     }
